refactor: ScoreString관련 폴더 이동및 필드 영역수정
